requirements.txt
blinker==1.8.1
click==8.1.7
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
slack-bolt==1.18.1
slack_sdk==3.27.1
Werkzeug==3.0.2
zipp==3.18.1